+++ src/metaSMT/backend/STP.hpp
@@ -223,10 +223,10 @@ namespace metaSMT {
       }
 
       result_type operator()( bvtags::bvsint_tag , boost::any arg ) {
-        typedef boost::tuple<int64_t, uint64_t> Tuple;
+        typedef boost::tuple<int64_t, unsigned> Tuple;
         Tuple tuple = boost::any_cast<Tuple>(arg);
         int64_t value = boost::get<0>(tuple);
-        uint64_t width = boost::get<1>(tuple);
+        unsigned width = boost::get<1>(tuple);
 
         if ( width > 8*sizeof(uint64_t)
              || value > std::numeric_limits<int64_t>::max()
